refactor(server): move dashboard rendering into a dedicated gui module
Extracts the templates, render helpers, and `serve_dashboard` handler
from `handlers.rs` into a new `gui.rs`. `handlers.rs` is now the home
for server-wide endpoints only (just `/health` today). No behavior
change.
